finalize styling for table =
